Move SimpleFileStorage to services
